src-srv/server.c updated: fix probleme équipe de type 1

+++ src-srv/server.c
@@ -32,7 +32,7 @@ int main(int argc, char **args) {
     client_t client = {0};
     if (accept_client(&client) == 0) {
       // Si le client est accepté, on récupère le message du client
-      uint8_t message;
+      uint16_t message;
       recv(client.sock, &message, sizeof(message), 0);
 
       // On décode son message
